Define the minimum hashing algorithm to be used when signing the CSR.
A list of deprecated hashing algorithms, that should not be used as
they are considered insecure.
Definition of the minimum hashing algorithm of SHA-2 to be used,
with the limitation that the latest guidance on recommended
hashing algorithms should be followed.

Did not include a mention of AEAD due to its limited adoption