Add mitigation for low power device renewals
Some devices are unable to perform a certificate renewal without
affecting the primary operation of the device. To prevent these
devices from performing a renewal while in use, a stop gap measure has been
proposed until a solution for low power devices is finalised.

The proposal is to disable EST after a successful enrolments,
allowing a system administrator to choose an appropriate time for the
device to perform the renewal by enabling EST again manually.