Replace recommended lifetime of client cert with period of 30 years
Remove the current recommended lifetime of 99991231235959Z with a
with a minimum of 30 years. This is due to issue with generating
CA's, intermediates and clients certs with the a notAfter date of
99991231235959Z and lifetime of products unlikely to be over 30
years.

Highlights that the Client certificate's chain of trust must
also be valid for the lifetime of the certifcate for the
chain to valid.

Makes it clear that EST must check revocation status of
TLS Client Certifcates.

Closes #6